feat: FEAT-161 live-browser UI audit — fix ai-ecosystem 500 + dashboard loading/empty states
Logged into the deployed app as an end user via Playwright and walked
every screen. Confirmed and fixed two real, live bugs:

- GET /api/v1/ai-ecosystem/metrics returned HTTP 500 (Postgres refuses a
  direct bool->double cast; cast to Integer first instead), silently
  swallowed by the frontend.
- Every dashboard widget rendered "loading" (null) and "genuinely empty"
  ([]) identically ("0 waiting", "0 open", literal "Loading..." in the
  real heading style), confirmed live on desktop and mobile. Added a
  shared skeleton/empty-state pair and wired it into all 9 widgets.

Verified: tsc clean, 40/40 test files / 295/295 tests passing (9 new).

+++ backend/src/api/v1/ai_ecosystem.py
@@ -14,7 +14,7 @@ from datetime import datetime, timedelta, timezone
 from typing import Literal
 
 from fastapi import APIRouter, Depends, Query
-from sqlalchemy import Float, cast, func, select
+from sqlalchemy import cast, func, select
 from sqlalchemy.ext.asyncio import AsyncSession
 from src.auth.dependencies import get_current_user
 from src.models.ai_ecosystem import AgentRegistry, AgentUsageLog
@@ -107,7 +107,9 @@ async def get_metrics(
                 func.coalesce(func.avg(AgentUsageLog.tokens_used), 0).label(
                     "avg_tokens"
                 ),
-                func.avg(cast(AgentUsageLog.success, Float)).label("success_rate"),
+                # Postgres/asyncpg refuses a direct bool->double cast
+                # (CannotCoerceError); int->double is the standard detour.
+                func.avg(cast(AgentUsageLog.success, Integer)).label("success_rate"),
             )
             .where(AgentUsageLog.invoked_at >= since)
             .group_by(AgentUsageLog.agent_name)
